Moved `PersonProperty` and related macros to submodule `people::property`.
