+++ src/tabulator.rs
@@ -4,7 +4,8 @@ A `Tabulator` is a mechanism by which reports know their columns.
 
 */
 
-use crate::{Context, ContextPeopleExt, PersonProperty};
+use crate::people::PersonProperty;
+use crate::{Context, ContextPeopleExt};
 use seq_macro::seq;
 use std::any::TypeId;
